# lib/api-zod/src/generated/types/departmentUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Staff Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface DepartmentUpdate {
  /**
     * @minLength 1
     * @maxLength 100
     */
  name?: string;
}