# lib/api-zod/src/generated/types/userWithRoles.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Staff Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { Role } from './role';

export interface UserWithRoles {
  id: number;
  full_name: string;
  username: string;
  email: string;
  /** @nullable */
  department_id?: number | null;
  /** @nullable */
  department_name?: string | null;
  /** @nullable */
  signature_image?: string | null;
  created_at: Date;
  updated_at: Date;
  roles: Role[];
}